Correct invalid keywords.txt KEYWORD_TOKENTYPE
The DATA_TYPE keyword identifier is only valid when used in the RSYNTAXTEXTAREA_TOKENTYPE field of keywords.txt. That is the fourth field rather than the second, where DATA_TYPE was previously. I moved DATA_TYPE to the fourth field and replaced it with the equivalent KEYWORD1 KEYWORD_TOKENTYPE in the second field.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#keywords

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#keyword_tokentype